Temporary move mdconf
